user_services/pyproject.toml
[tool.poetry]
name = "user-services"
version = "0.1.0"
description = ""
authors = ["Azeem000011111111 <139073260+Azeem000011111111@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
sqlmodel = "^0.0.18"
aiokafka = "^0.10.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
psycopg = {extras = ["binary"], version = "^3.1.18"}
passlib = "^1.7.4"
alembic = "^1.13.1"
bcrypt = "^4.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
